lib/ogsf, lib/nviz: Fix multiplications converted to a larger type
Cast the first operand so that offset computations like
row * ymod * cols are done in long, not int, matching the long
variables they are assigned to, and so that float products assigned
to double variables are computed in double. In gsd_surf2model, make
the ternary literals float to avoid a pointless float-double-float
round trip for values stored back into a float Point3.

Fixes 33 CodeQL cpp/integer-multiplication-cast-to-long alerts.

+++ lib/ogsf/gsd_cplane.c
@@ -67,9 +67,9 @@ void gsd_def_cplane(int num, float *pt, float *norm)
     ppt[1] = (pt[1] + Cp_pt[1]) * sy;
     ppt[2] = (pt[2] + Cp_pt[2] - zmin) * sz;
 
-    params[0] = norm[0] * sx;
-    params[1] = norm[1] * sy;
-    params[2] = norm[2] * sz;
+    params[0] = (double)norm[0] * sx;
+    params[1] = (double)norm[1] * sy;
+    params[2] = (double)norm[2] * sz;
     GS_dv3norm(params);
     params[3] = -ppt[0] * params[0] - ppt[1] * params[1] - ppt[2] * params[2];
